Split invitation view into separate component
