+++ src/app/invitations/invitations.component.html
@@ -4,106 +4,7 @@
 
 <ng-container *ngrxLet="payload$ as payload; suspenseTpl: loading; error as err">
   <ng-container *ngIf="payload">
-    <ng-container *ngFor="let vm of payload.invitations; index as i">
-      <div class="surface-card p-4 shadow-2 border-round mb-4">
-        <div class="flex flex-row justify-content-between">
-          <div class="text-3xl font-medium text-900 mb-3">
-            {{ vm.model.spec.email }}
-          </div>
-        </div>
-        <div class="flex flex-wrap border-top-1 surface-border">
-          <ul class="w-full list-none p-0 m-0 surface-border">
-            <li class="flex align-items-center py-2 px-2 flex-wrap surface-ground">
-              <div class="text-500 w-full md:w-3 font-medium" i18n>ID</div>
-              <div class="text-900 w-full md:w-9">
-                {{ vm.model.metadata.name }}
-              </div>
-            </li>
-            <li class="flex align-items-center py-2 px-2 flex-wrap surface-ground">
-              <div class="text-500 w-full md:w-3 font-medium" i18n>Note</div>
-              <div class="text-900 w-full md:w-9">
-                {{ vm.model.spec.note }}
-              </div>
-            </li>
-
-            <li *ngIf="vm.model.status?.validUntil" class="flex align-items-center py-2 px-2 flex-wrap surface-ground">
-              <div class="text-500 w-full md:w-3 font-medium" i18n>{{ vm.hasExpired ? 'Expired' : 'Expires' }}</div>
-              <div class="text-900 w-full md:w-9">
-                {{ vm.expires }}
-              </div>
-            </li>
-
-            <li class="flex align-items-center py-2 px-2 flex-wrap surface-ground">
-              <div class="text-500 w-full md:w-3 font-medium" i18n>Status</div>
-              <div class="text-900 w-full md:w-9">
-                <p-tag
-                  *ngFor="let condition of vm.model.status?.conditions"
-                  [rounded]="true"
-                  [severity]="severityOfCondition(condition)"
-                  [value]="condition.message ?? ''"
-                  styleClass="mr-2"/>
-                <p-tag
-                  *ngIf="vm.hasExpired"
-                  [rounded]="true"
-                  severity="danger"
-                  value="Expired"
-                  styleClass="mr-2"/>
-                <p-tag *ngIf="!vm.model.status?.conditions"
-                       [rounded]="true"
-                       severity="info"
-                       value="Pending"
-                       styleClass="mr-2"/>
-              </div>
-            </li>
-
-
-            <p-table [value]="vm.permissionTable" styleClass="p-datatable-sm">
-              <ng-template pTemplate="header">
-                <tr>
-                  <th class="p-2" i18n>Name</th>
-                  <th i18n>Viewer</th>
-                  <th i18n>Admin</th>
-                  <th i18n>Teams</th>
-                </tr>
-              </ng-template>
-              <ng-template pTemplate="body" let-permission>
-                <tr>
-                  <td>
-                    <a [routerLink]="['/', permission.kind, permission.slug]"
-                       class="text-blue-500 hover:text-primary cursor-pointer no-underline" [title]="permission.slug">
-                      <span class="pr-1 white-space-nowrap">
-                        <ng-container [ngSwitch]="permission.kind">
-                          <fa-icon *ngSwitchCase="'organizations'" [icon]="faSitemap" [fixedWidth]="true" />
-                          <fa-icon *ngSwitchCase="'billingentities'" [icon]="faDollarSign" [fixedWidth]="true" />
-                        </ng-container>
-                        {{ permission.displayName ? permission.displayName : permission.slug }}
-                      </span>
-                    </a>
-                  </td>
-                  <td>
-                    <fa-icon *ngIf="permission.viewer" [icon]="faCheck"/>
-                  </td>
-                  <td>
-                    <fa-icon *ngIf="permission.admin" [icon]="faCheck"/>
-                  </td>
-                  <td>
-                    <ng-container *ngFor="let team of permission.teams">
-                      <a [routerLink]="['/teams', permission.slug, team.slug]"
-                         class="text-blue-500 hover:text-primary cursor-pointer no-underline m-2" title="{{ permission.slug }}/{{ team.slug }}">
-                        <span class="pr-1 white-space-nowrap">
-                          <fa-icon [icon]="faUserGroup"/>
-                          {{ team.displayName ? team.displayName : team.slug }}
-                        </span>
-                      </a>
-                    </ng-container>
-                  </td>
-                </tr>
-              </ng-template>
-            </p-table>
-          </ul>
-        </div>
-      </div>
-    </ng-container>
+      <app-invitation-detail *ngFor="let invitation of payload.invitations" [invitation]="invitation" />
   </ng-container>
 
   <ng-container *ngIf="payload?.invitations?.length === 0 && !err">
